EDU-295 refactor: 라우터 정리
