+++ app/routers/activity_log_router.py
@@ -1,4 +1,4 @@
-# app/routers/activity_log.py
+# app/routers/activity_log_router.py
 
 from fastapi import APIRouter
